+++ src/js/utils.js
@@ -152,8 +152,6 @@ function readCookie( name ) {
 
         return null;
     }
-
-
 }
 
 /**
@@ -236,6 +234,37 @@ function resizeImage( file, maxPixels ) {
     } );
 }
 
+/**
+ * Copied from: https://gist.github.com/creationix/7435851
+ * Joins path segments.  Preserves initial "/" and resolves ".." and "."
+ * Does not support using ".." to go above/outside the root.
+ * This means that join("foo", "../../bar") will not resolve to "../bar"
+ */
+function joinPath( /* path segments */ ) {
+    // Split the inputs into a list of path commands.
+    let parts = [];
+    for ( var i = 0, l = arguments.length; i < l; i++ ) {
+        parts = parts.concat( arguments[i].split( '/' ) );
+    }
+    // Interpret the path commands to get the new resolved path.
+    let newParts = [];
+    for ( i = 0, l = parts.length; i < l; i++ ) {
+        var part = parts[i];
+        // Remove leading and trailing slashes
+        // Also remove "." segments
+        if ( !part || part === '.' ) continue;
+        // Interpret ".." to pop the last segment
+        if ( part === '..' ) newParts.pop();
+        // Push new path segments.
+        else newParts.push( part );
+    }
+    // Preserve the initial slash if there was one.
+    if ( parts[0] === '' ) newParts.unshift( '' );
+
+    // Turn back into a single string path.
+    return newParts.join( '/' ) || ( newParts.length ? '/' : '.' );
+}
+
 export {
     parseFunctionFromExpression,
     stripQuotes,
@@ -245,5 +274,6 @@ export {
     readCookie,
     dataUriToBlobSync,
     getPasteData,
-    resizeImage
+    resizeImage,
+    joinPath,
 };